Move weather mocks
