--- tests/e2e/mocks/index.js
@@ -1,4 +1,0 @@
-const generateWeather = require("./weather_current");
-const generateWeatherForecast = require("./weather_forecast");
-
-module.exports = { generateWeather, generateWeatherForecast };
